fix(kfd): 修 kfd_helper 编译剩余两处错误
CI run 36209447623 kfd_helper 步骤日志:
1. kfd_helper.c:204/211/216 use of undeclared CSMAGIC_EMBEDDED_SIGNATURE/
   CSSLOT_CODEDIRECTORY/CS_HASHTYPE_* -> 补 #include <mach-o/codesign.h>
   (cs_superblob/cs_blobindex/cs_codedirectory 结构由该头提供)
2. kfd_helper.c:319 kclose(kfd) passing 'struct kfd*' to u64 ->
   kclose((u64)kfd) (libkfd 公开 API 参数为 u64 句柄)

注: libkfd kopen 返回 (u64)指针,kfd_helper (struct kfd*)kopen 恢复指针正确;
kread((u64)k,..) 已强转。至此 kfd_helper 应可编入 IPA。

+++ tools/kfd_helper.c
@@ -41,6 +41,7 @@
 #include <sys/stat.h>
 #include <mach-o/loader.h>
 #include <mach-o/fat.h>
+#include <mach-o/codesign.h>   /* CSMAGIC_EMBEDDED_SIGNATURE, cs_superblob, csslot_codedirectory */
 #include <CommonCrypto/CommonDigest.h>   // CC_SHA256 — 算 cdhash
 
 /* libkfd 公开 API（以你 clone 的版本为准，这里按通用签名占位）
@@ -316,7 +317,7 @@ int main(int argc, char **argv) {
 
     int rc = inject_trust_cache(kfd, cdhash);
 
-    kclose(kfd);
+    kclose((u64)kfd);
     fprintf(stderr, "[kfd-helper] %s\n", rc == 0 ? "ok" : "failed");
     return rc == 0 ? 0 : 1;
 }
